+++ gecode/set/array.cpp
@@ -61,8 +61,10 @@ namespace Gecode {
     Set::Limits::check(ubMin,"SetVarArray::SetVarArray");
     Set::Limits::check(ubMax,"SetVarArray::SetVarArray");
     Set::Limits::check(maxCard,"SetVarArray::SetVarArray");
-    unsigned int glbSize = (lbMin <= lbMax ? lbMax-lbMin+1 : 0);
-    unsigned int lubSize = (ubMin <= ubMax ? ubMax-ubMin+1 : 0);
+    unsigned int glbSize = 
+      (lbMin <= lbMax ? static_cast<unsigned int>(lbMax-lbMin+1) : 0U);
+    unsigned int lubSize = 
+      (ubMin <= ubMax ? static_cast<unsigned int>(ubMax-ubMin+1) : 0U);
     if (minCard > maxCard || minCard > lubSize || maxCard < glbSize ||
         lbMin < ubMin || lbMax > ubMax)
       throw Set::VariableEmptyDomain("SetVarArray::SetVarArray");
@@ -80,7 +82,8 @@ namespace Gecode {
     Set::Limits::check(maxCard,"SetVarArray::SetVarArray");
     IntSetRanges glbr(glb);
     unsigned int glbSize = Iter::Ranges::size(glbr);
-    unsigned int lubSize = (ubMin <= ubMax ? ubMax-ubMin+1 : 0);
+    unsigned int lubSize = 
+      (ubMin <= ubMax ? static_cast<unsigned int>(ubMax-ubMin+1) : 0U);
     if (minCard > maxCard || minCard > lubSize || maxCard < glbSize ||
         glb.min() < ubMin || glb.max() > ubMax)
       throw Set::VariableEmptyDomain("SetVarArray::SetVarArray");
@@ -99,7 +102,8 @@ namespace Gecode {
     Iter::Ranges::Singleton glbr(lbMin,lbMax);
     IntSetRanges lubr(lub);
     IntSetRanges lubr_s(lub);
-    unsigned int glbSize = (lbMin <= lbMax ? lbMax-lbMin+1 : 0);
+    unsigned int glbSize = 
+      (lbMin <= lbMax ? static_cast<unsigned int>(lbMax-lbMin+1) : 0U);
     unsigned int lubSize = Iter::Ranges::size(lubr_s);
     if (minCard > maxCard || minCard > lubSize || maxCard < glbSize ||
         !Iter::Ranges::subset(glbr,lubr))
@@ -138,8 +142,10 @@ namespace Gecode {
     Set::Limits::check(ubMin,"SetVarArgs::SetVarArgs");
     Set::Limits::check(ubMax,"SetVarArgs::SetVarArgs");
     Set::Limits::check(maxCard,"SetVarArgs::SetVarArgs");
-    unsigned int glbSize = (lbMin <= lbMax ? lbMax-lbMin+1 : 0);
-    unsigned int lubSize = (ubMin <= ubMax ? ubMax-ubMin+1 : 0);
+    unsigned int glbSize = 
+      (lbMin <= lbMax ? static_cast<unsigned int>(lbMax-lbMin+1) : 0U);
+    unsigned int lubSize = 
+      (ubMin <= ubMax ? static_cast<unsigned int>(ubMax-ubMin+1) : 0U);
     if (minCard > maxCard || minCard > lubSize || maxCard < glbSize ||
         lbMin < ubMin || lbMax > ubMax)
       throw Set::VariableEmptyDomain("SetVarArgs::SetVarArgs");
@@ -157,7 +163,8 @@ namespace Gecode {
     Set::Limits::check(maxCard,"SetVarArgs::SetVarArgs");
     IntSetRanges glbr(glb);
     unsigned int glbSize = Iter::Ranges::size(glbr);
-    unsigned int lubSize = (ubMin <= ubMax ? ubMax-ubMin+1 : 0);
+    unsigned int lubSize = 
+      (ubMin <= ubMax ? static_cast<unsigned int>(ubMax-ubMin+1) : 0U);
     if (minCard > maxCard || minCard > lubSize || maxCard < glbSize ||
         glb.min() < ubMin || glb.max() > ubMax)
       throw Set::VariableEmptyDomain("SetVarArgs::SetVarArgs");
@@ -176,7 +183,8 @@ namespace Gecode {
     Iter::Ranges::Singleton glbr(lbMin,lbMax);
     IntSetRanges lubr(lub);
     IntSetRanges lubr_s(lub);
-    unsigned int glbSize = (lbMin <= lbMax ? lbMax-lbMin+1 : 0);
+    unsigned int glbSize = 
+      (lbMin <= lbMax ? static_cast<unsigned int>(lbMax-lbMin+1) : 0U);
     unsigned int lubSize = Iter::Ranges::size(lubr_s);
     if (minCard > maxCard || minCard > lubSize || maxCard < glbSize ||
         !Iter::Ranges::subset(glbr,lubr))
